pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "img-conv"
version = "0.1.0"
description = "A powerful command-line tool for batch image processing and format conversion"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: System :: Archiving :: Compression",
    "Topic :: Utilities",
]
keywords = ["image", "conversion", "webp", "png", "jpeg", "cli", "batch-processing"]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/image-converter"
Repository = "https://github.com/yourusername/image-converter"
Documentation = "https://github.com/yourusername/image-converter/blob/main/docs/"
Issues = "https://github.com/yourusername/image-converter/issues"

[project.scripts]
img-conv = "img_conv.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/img_conv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''